backend/requirements.txt
fastapi==0.115.12
psycopg2-binary==2.9.9
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
pydantic==2.11.7
pydantic-settings==2.9.1
google-genai>=1.70.0
scikit-learn==1.6.1
reportlab==4.4.1
Pillow==11.2.1
python-multipart==0.0.20
python-dotenv==1.1.1
aiofiles==24.1.0
httpx==0.28.1
pypdf==5.4.0
